requirements.txt
Faker==37.3.0
pymongo==4.13.1
requests==2.32.3
robotframework==7.2.2
robotframework-browser==19.5.1
robotframework-faker==5.0.0
robotframework-requests==0.9.7
